Create a tab for Adding Events, next to the Agenda tab.
It's a single place for adding events manually (not yet implemented)
and for adding events from the page. Also simplifies the rest of the
logic in displaying the list of events: there is no longer a separate
rendering for a single event (it appears as a list of events containing
one event), and the Add tab is always shown, making place for manually
adding events.