Improvements to Dutch translations
Dutch translations were not consistent with the Google Calendar UI. Also, the word "Evenement" in Dutch does not really map 1-on-1 to "Event" in english. An "evenement" has connotations of something being organized for a large group of people, like a concert or festival.

I've replaced "evenement" with either "gebeurtenis" (which maps more closely to the english "event") when it concerns something on the page you're looking at, or afspraak ("appointment") when it concerns something that is in your calendar.