Add document icon for quickly accessing event attachments
Make it possibly to quickly open a Google Doc associated with some event
by adding an icon to the event entry, à la the location and Hangout icons.

If there are multiple attachments, do nothing. This ideally would have a
nice UI, but I think it's OK because multiple attachments are the
exception.

Credit to mrmcd@ for the idea.